+++ ObjectPosition.h
@@ -13,13 +13,13 @@ public:
 	virtual ~ObjectPosition(){};
 
 public:
-	cv::Point2i fieldCoords; // (x, y) Coordinates to display objects on field by, relative to field
+	cv::Point2d fieldCoords; // (x, y) Coordinates to display objects on field by, relative to field
 	cv::Point2i rawPixelCoords; // (x, y) Raw from frame
 	cv::Point2d polarMetricCoords;      // (distance, angle) Relative to robot
 	virtual void updateFieldCoords(cv::Point orgin = cv::Point(0, 0)) = 0;
 
 protected:
-	cv::Point2i lastFieldCoords;
+	cv::Point2d lastFieldCoords;
 };
